@@ -2,17 +2,19 @@ apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'com.google.gms.google-services'
 apply from: '../config/quality/quality.gradle'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
     defaultConfig {
         applicationId "org.sagebionetworks.research.mpower"
         minSdkVersion 19
-        targetSdkVersion 27
-        versionCode 13
-        versionName "0.1.0"
+        targetSdkVersion 28
+        // this is not the final version code, product flavors add to it
+        versionCode 10
+        versionName "0.2.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
         vectorDrawables.useSupportLibrary = true
@@ -28,6 +30,7 @@ android {
     buildTypes {
         debug {
             minifyEnabled false
+            versionNameSuffix "-debug"
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             testProguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguardTest-rules.pro'
         }
@@ -42,9 +45,12 @@ android {
 
     productFlavors {
         internal {
+            versionCode 1 + 2 * android.defaultConfig.versionCode
+            versionNameSuffix "-internal"
             dimension "userRole"
         }
         participant {
+            versionCode 2 * android.defaultConfig.versionCode
             dimension "userRole"
         }
     }
@@ -112,9 +118,9 @@ dependencies {
     // room db functionality
     implementation 'android.arch.persistence.room:runtime:1.1.1'
     kapt 'android.arch.persistence.room:compiler:1.1.1'
-    androidTestImplementation'android.arch.persistence.room:testing:1.1.1'
+    androidTestImplementation 'android.arch.persistence.room:testing:1.1.1'
 
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
 
     implementation 'nl.littlerobots.rxlint:rxlint:1.6.1'
 
@@ -122,9 +128,17 @@ dependencies {
 
     testImplementation 'junit:junit:4.12'
 
-    androidTestImplementation  'org.mockito:mockito-android:2.21.0'
+    androidTestImplementation 'org.mockito:mockito-android:2.21.0'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
     androidTestImplementation 'com.android.support.test:rules:1.0.2'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'org.hamcrest:hamcrest-library:1.3'
+
+    // firebase
+    implementation 'com.google.firebase:firebase-core:16.0.3'
+}
+tasks.withType(JavaCompile) {
+    options.compilerArgs += [
+            '-XepExcludedPaths:.*/build/generated/.*'
+    ]
 }
